+++ openmls/src/treesync/mod.rs
@@ -53,7 +53,7 @@ use crate::{
     ciphersuite::{signable::Verifiable, Secret},
     credentials::CredentialWithKey,
     error::LibraryError,
-    extensions::Extensions,
+    extensions::{Extensions, RequiredCapabilitiesExtension},
     group::{GroupId, Member},
     key_packages::Lifetime,
     messages::{PathSecret, PathSecretError},
@@ -82,8 +82,8 @@ pub use node::encryption_keys::EncryptionKey;
 // Public re-exports
 pub use node::{
     leaf_node::{
-        LeafNode, LeafNodeParameters, LeafNodeParametersBuilder, LeafNodeSource,
-        LeafNodeUpdateError,
+        CapabilitiesPolicy, LeafNode, LeafNodeBuildError, LeafNodeParameters,
+        LeafNodeParametersBuilder, LeafNodeSource, LeafNodeUpdateError,
     },
     parent_node::ParentNode,
     Node,
@@ -440,6 +440,7 @@ impl TreeSync {
     ///
     /// Returns the resulting [`TreeSync`] instance, as well as the
     /// corresponding [`CommitSecret`].
+    #[allow(clippy::too_many_arguments)]
     pub(crate) fn new(
         provider: &impl OpenMlsProvider,
         signer: &impl Signer,
@@ -448,7 +449,9 @@ impl TreeSync {
         life_time: Lifetime,
         capabilities: Capabilities,
         extensions: Extensions<LeafNode>,
-    ) -> Result<(Self, CommitSecret, EncryptionKeyPair), LibraryError> {
+        required_capabilities: Option<RequiredCapabilitiesExtension>,
+        capabilities_policy: CapabilitiesPolicy,
+    ) -> Result<(Self, CommitSecret, EncryptionKeyPair), LeafNodeBuildError> {
         let new_leaf_node_params = NewLeafNodeParams {
             ciphersuite,
             credential_with_key,
@@ -457,6 +460,8 @@ impl TreeSync {
             capabilities,
             extensions,
             tree_info_tbs: TreeInfoTbs::KeyPackage,
+            required_capabilities,
+            capabilities_policy,
         };
         let (leaf, encryption_key_pair) = LeafNode::new(provider, signer, new_leaf_node_params)?;
 
@@ -501,7 +506,9 @@ impl TreeSync {
         capabilities: Capabilities,
         leaf_extensions: Extensions<LeafNode>,
         encryption_key_pair: EncryptionKeyPair,
-    ) -> Result<(Self, EncryptionKeyPair), LibraryError> {
+        required_capabilities: Option<RequiredCapabilitiesExtension>,
+        capabilities_policy: CapabilitiesPolicy,
+    ) -> Result<(Self, EncryptionKeyPair), LeafNodeBuildError> {
         let new_leaf_node_params = NewLeafNodeParams {
             ciphersuite,
             credential_with_key,
@@ -511,6 +518,8 @@ impl TreeSync {
             capabilities,
             extensions: leaf_extensions,
             tree_info_tbs: TreeInfoTbs::KeyPackage,
+            required_capabilities,
+            capabilities_policy,
         };
         let (leaf, encryption_key_pair) = LeafNode::new_with_encryption_key_pair(
             signer,